+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added "Set Up renv" button to initialize the environment, install
+  dependencies, and create a lockfile when no valid R lockfile file is present.
+
 ## [1.21.0]
 
 ### Added
@@ -19,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed an issue deploying content to free Posit Connect Cloud accounts with incorrect  
+- Fixed an issue deploying content to free Posit Connect Cloud accounts with incorrect
   access settings. (#3005)
 
 ## [1.19.1]